fix: re-apply DashboardEditActionEvent extraction for circular dependency
Re-applies the circular dependency fix to pass CI checks.
Extracts DashboardEditActionEvent to a separate file to break the cycle:
shared.ts -> VizPanelEditableElement.tsx -> useQuickEditOptions.tsx -> shared.ts

Made-with: Cursor
